bugfixing editari

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Data/Entities/WorkflowHistorieLogs/WorkflowHistoryLog.cs
@@ -8,7 +8,7 @@ namespace DigitNow.Domain.DocumentManagement.Data.Entities
     {
         public long DocumentId { get; set; }
         public DocumentStatus DocumentStatus { get; set; }
-        public int DestinationDepartmentId { get; set; }
+        public long DestinationDepartmentId { get; set; }
         public int RecipientType { get; set; } 
         public long RecipientId { get; set; }
         public string RecipientName { get; set; }
